+++ packages/admin-ui/client/components/PageLoading.js
@@ -1,4 +1,5 @@
-import React from 'react';
+/** @jsx jsx */
+import { jsx } from '@emotion/core';
 import { LoadingIndicator } from '@voussoir/ui/src/primitives/loading';
 
 export default function PageLoading() {