+++ packages/fields/types/CalendarDay/views/Filter.js
@@ -11,22 +11,29 @@ type Props = {
   filter: Object,
   innerRef: Ref<*>,
   onChange: Event => void,
+  recalcHeight: () => void,
 };
 
-export default class CalendarDayFilterView extends Component<Props> {
-  constructor(props) {
+type State = {
+  value: string,
+};
+
+type SetDate = Date => void;
+
+export default class CalendarDayFilterView extends Component<Props, State> {
+  constructor(props: Props) {
     super(props);
     this.state = { value: format(new Date(), FORMAT) };
   }
 
-  handleDayClick = day => {
+  handleDayClick = (day: Date) => {
     const { onChange } = this.props;
     const value = format(day, FORMAT);
     onChange(value);
     this.setState({ value });
   };
 
-  handleMonthSelect = (event, setDate, setSelectedDate) => {
+  handleMonthSelect = (event: Object, setDate: SetDate, setSelectedDate: SetDate) => {
     const { field, onChange } = this.props;
     const month = event.target.value;
     const newDate = setMonth(this.state.value, month);
@@ -34,10 +41,10 @@ export default class CalendarDayFilterView extends Component<Props> {
     setDate(newDate);
     setSelectedDate(newDate);
     this.setState({ value });
-    onChange(field, value);
+    onChange(field);
   };
 
-  handleYearSelect = (event, setDate, setSelectedDate) => {
+  handleYearSelect = (event: Object, setDate: SetDate, setSelectedDate: SetDate) => {
     const { field, onChange } = this.props;
     const year = event.target.value;
     const newDate = setYear(this.state.value, year);
@@ -45,10 +52,10 @@ export default class CalendarDayFilterView extends Component<Props> {
     setDate(newDate);
     setSelectedDate(newDate);
     this.setState({ value });
-    onChange(field, value);
+    onChange(field);
   };
 
-  componentDidUpdate(prevProps) {
+  componentDidUpdate(prevProps: Props) {
     const { filter } = this.props;
 
     if (prevProps.filter !== filter) {