+++ packages/mongo-join-builder/executor.js
@@ -1,2 +1,0 @@
-module.exports = ({ joinQuery, mutator, aggregate }) =>
-  aggregate(joinQuery).then(queryResult => mutator(queryResult));