+++ packages/mongo-join-builder/tests/join-builder.test.js
@@ -11,11 +11,13 @@ describe('join builder', () => {
       }
 
     */
-    const { joinQuery } = joinBuilder({
-      pipeline: [{ name: { $eq: 'foobar' } }, { age: { $eq: 23 } }],
+    const { pipeline } = joinBuilder({
+      matchTerm: { $and: [{ name: { $eq: 'foobar' } }, { age: { $eq: 23 } }] },
+      postJoinPipeline: [],
+      relationships: {},
     });
 
-    expect(joinQuery).toMatchObject([
+    expect(pipeline).toMatchObject([
       { $match: { $and: [{ name: { $eq: 'foobar' } }, { age: { $eq: 23 } }] } },
       { $addFields: { id: '$_id' } },
     ]);
@@ -34,38 +36,37 @@ describe('join builder', () => {
       }
 
     */
-    const { joinQuery } = joinBuilder({
+    const { pipeline } = joinBuilder({
       relationships: {
         abc123: {
           from: 'user-collection',
           field: 'author',
-          pipeline: [{ name: { $eq: 'Alice' } }],
+          matchTerm: { name: { $eq: 'Alice' } },
+          postJoinPipeline: [],
           postQueryMutation: jest.fn(),
           many: false,
+          relationships: {},
         },
       },
-      pipeline: [
-        { title: { $eq: 'foobar' } },
-        { views: { $eq: 23 } },
-        { abc123_author_every: { $eq: true } },
-      ],
+      matchTerm: {
+        $and: [
+          { title: { $eq: 'foobar' } },
+          { views: { $eq: 23 } },
+          { abc123_author_every: { $eq: true } },
+        ],
+      },
+      postJoinPipeline: [],
     });
 
-    expect(joinQuery).toMatchObject([
+    expect(pipeline).toMatchObject([
       {
         $lookup: {
           from: 'user-collection',
           as: 'abc123_author',
           let: { abc123_author_id: '$author' },
           pipeline: [
-            {
-              $match: {
-                $and: [
-                  { $expr: { $eq: ['$_id', '$$abc123_author_id'] } },
-                  { name: { $eq: 'Alice' } },
-                ],
-              },
-            },
+            { $match: { $expr: { $eq: ['$_id', '$$abc123_author_id'] } } },
+            { $match: { name: { $eq: 'Alice' } } },
             { $addFields: { id: '$_id' } },
           ],
         },
@@ -101,24 +102,28 @@ describe('join builder', () => {
       }
 
     */
-    const { joinQuery } = joinBuilder({
+    const { pipeline } = joinBuilder({
       relationships: {
         abc123: {
           from: 'posts-collection',
           field: 'posts',
-          pipeline: [],
+          postJoinPipeline: [],
           postQueryMutation: jest.fn(),
           many: true,
+          relationships: {},
         },
       },
-      pipeline: [
-        { name: { $eq: 'foobar' } },
-        { age: { $eq: 23 } },
-        { abc123_posts_some: { $eq: true } },
-      ],
+      matchTerm: {
+        $and: [
+          { name: { $eq: 'foobar' } },
+          { age: { $eq: 23 } },
+          { abc123_posts_some: { $eq: true } },
+        ],
+      },
+      postJoinPipeline: [],
     });
 
-    expect(joinQuery).toMatchObject([
+    expect(pipeline).toMatchObject([
       {
         $lookup: {
           from: 'posts-collection',
@@ -161,26 +166,28 @@ describe('join builder', () => {
       }
 
     */
-    const { joinQuery } = joinBuilder({
+    const { pipeline } = joinBuilder({
       relationships: {
         abc123: {
           from: 'posts-collection',
           field: 'posts',
-          pipeline: [],
           postJoinPipeline: [{ $orderBy: 'title' }],
           postQueryMutation: jest.fn(),
           many: true,
+          relationships: {},
         },
       },
-      pipeline: [
-        { name: { $eq: 'foobar' } },
-        { age: { $eq: 23 } },
-        { abc123_posts_some: { $eq: true } },
-      ],
+      matchTerm: {
+        $and: [
+          { name: { $eq: 'foobar' } },
+          { age: { $eq: 23 } },
+          { abc123_posts_some: { $eq: true } },
+        ],
+      },
       postJoinPipeline: [{ $limit: 10 }],
     });
 
-    expect(joinQuery).toMatchObject([
+    expect(pipeline).toMatchObject([
       {
         $lookup: {
           from: 'posts-collection',
@@ -233,68 +240,73 @@ describe('join builder', () => {
       }
 
     */
-    const { joinQuery } = joinBuilder({
+    const { pipeline } = joinBuilder({
       relationships: {
         abc123: {
           from: 'posts-collection',
           field: 'posts',
-          pipeline: [{ title: { $eq: 'hello' } }, { def456_tags_some: { $eq: true } }],
+          matchTerm: { $and: [{ title: { $eq: 'hello' } }, { def456_tags_some: { $eq: true } }] },
+          postJoinPipeline: [],
           postQueryMutation: jest.fn(),
           many: true,
           relationships: {
             def456: {
               from: 'tags-collection',
               field: 'tags',
-              pipeline: [{ name: { $eq: 'React' } }, { xyz890_posts_every: { $eq: true } }],
+              matchTerm: {
+                $and: [{ name: { $eq: 'React' } }, { xyz890_posts_every: { $eq: true } }],
+              },
+              postJoinPipeline: [],
               postQueryMutation: jest.fn(),
               many: true,
               relationships: {
                 xyz890: {
                   from: 'posts-collection',
                   field: 'posts',
-                  pipeline: [{ published: { $eq: true } }],
+                  matchTerm: { published: { $eq: true } },
+                  postJoinPipeline: [],
                   postQueryMutation: jest.fn(),
                   many: true,
+                  relationships: {},
                 },
               },
             },
           },
         },
       },
-      pipeline: [
-        { name: { $eq: 'foobar' } },
-        { age: { $eq: 23 } },
-        { abc123_posts_some: { $eq: true } },
-      ],
+      matchTerm: {
+        $and: [
+          { name: { $eq: 'foobar' } },
+          { age: { $eq: 23 } },
+          { abc123_posts_some: { $eq: true } },
+        ],
+      },
+      postJoinPipeline: [],
     });
 
-    expect(joinQuery).toMatchObject([
+    expect(pipeline).toMatchObject([
       {
         $lookup: {
           from: 'posts-collection',
           as: 'abc123_posts',
           let: { abc123_posts_ids: '$posts' },
           pipeline: [
+            { $match: { $expr: { $in: ['$_id', '$$abc123_posts_ids'] } } },
             {
               $lookup: {
                 from: 'tags-collection',
                 as: 'def456_tags',
                 let: { def456_tags_ids: '$tags' },
                 pipeline: [
+                  { $match: { $expr: { $in: ['$_id', '$$def456_tags_ids'] } } },
                   {
                     $lookup: {
                       from: 'posts-collection',
                       as: 'xyz890_posts',
                       let: { xyz890_posts_ids: '$posts' },
                       pipeline: [
-                        {
-                          $match: {
-                            $and: [
-                              { $expr: { $in: ['$_id', '$$xyz890_posts_ids'] } },
-                              { published: { $eq: true } },
-                            ],
-                          },
-                        },
+                        { $match: { $expr: { $in: ['$_id', '$$xyz890_posts_ids'] } } },
+                        { $match: { published: { $eq: true } } },
                         { $addFields: { id: '$_id' } },
                       ],
                     },
@@ -310,11 +322,7 @@ describe('join builder', () => {
                   },
                   {
                     $match: {
-                      $and: [
-                        { $expr: { $in: ['$_id', '$$def456_tags_ids'] } },
-                        { name: { $eq: 'React' } },
-                        { xyz890_posts_every: { $eq: true } },
-                      ],
+                      $and: [{ name: { $eq: 'React' } }, { xyz890_posts_every: { $eq: true } }],
                     },
                   },
                   { $addFields: { id: '$_id' } },
@@ -329,13 +337,7 @@ describe('join builder', () => {
               },
             },
             {
-              $match: {
-                $and: [
-                  { $expr: { $in: ['$_id', '$$abc123_posts_ids'] } },
-                  { title: { $eq: 'hello' } },
-                  { def456_tags_some: { $eq: true } },
-                ],
-              },
+              $match: { $and: [{ title: { $eq: 'hello' } }, { def456_tags_some: { $eq: true } }] },
             },
             { $addFields: { id: '$_id' } },
           ],
@@ -378,57 +380,56 @@ describe('join builder', () => {
       }
     */
 
-    const { joinQuery } = joinBuilder({
+    const { pipeline } = joinBuilder({
       relationships: {
         zip567: {
           from: 'posts-collection',
           field: 'posts',
-          pipeline: [{ title: { $eq: 'hello' } }, { quux987_labels_some: { $eq: true } }],
+          matchTerm: {
+            $and: [{ title: { $eq: 'hello' } }, { quux987_labels_some: { $eq: true } }],
+          },
+          postJoinPipeline: [],
           postQueryMutation: jest.fn(),
           many: true,
           relationships: {
             quux987: {
               from: 'labels-collection',
               field: 'labels',
-              pipeline: [{ name: { $eq: 'foo' } }],
+              matchTerm: { name: { $eq: 'foo' } },
+              postJoinPipeline: [],
               postQueryMutation: jest.fn(),
               many: true,
+              relationships: {},
             },
           },
         },
       },
-      pipeline: [
-        {
-          $and: [
-            { name: { $eq: 'foobar' } },
-            { age: { $eq: 23 } },
-            { zip567_posts_every: { $eq: true } },
-          ],
-        },
-      ],
+      matchTerm: {
+        $and: [
+          { name: { $eq: 'foobar' } },
+          { age: { $eq: 23 } },
+          { zip567_posts_every: { $eq: true } },
+        ],
+      },
+      postJoinPipeline: [],
     });
 
-    expect(joinQuery).toMatchObject([
+    expect(pipeline).toMatchObject([
       {
         $lookup: {
           from: 'posts-collection',
           as: 'zip567_posts',
           let: { zip567_posts_ids: '$posts' },
           pipeline: [
+            { $match: { $expr: { $in: ['$_id', '$$zip567_posts_ids'] } } },
             {
               $lookup: {
                 from: 'labels-collection',
                 as: 'quux987_labels',
                 let: { quux987_labels_ids: '$labels' },
                 pipeline: [
-                  {
-                    $match: {
-                      $and: [
-                        { $expr: { $in: ['$_id', '$$quux987_labels_ids'] } },
-                        { name: { $eq: 'foo' } },
-                      ],
-                    },
-                  },
+                  { $match: { $expr: { $in: ['$_id', '$$quux987_labels_ids'] } } },
+                  { $match: { name: { $eq: 'foo' } } },
                   { $addFields: { id: '$_id' } },
                 ],
               },
@@ -442,11 +443,7 @@ describe('join builder', () => {
             },
             {
               $match: {
-                $and: [
-                  { $expr: { $in: ['$_id', '$$zip567_posts_ids'] } },
-                  { title: { $eq: 'hello' } },
-                  { quux987_labels_some: { $eq: true } },
-                ],
+                $and: [{ title: { $eq: 'hello' } }, { quux987_labels_some: { $eq: true } }],
               },
             },
             { $addFields: { id: '$_id' } },
@@ -490,58 +487,55 @@ describe('join builder', () => {
       }
     */
 
-    const { joinQuery } = joinBuilder({
+    const { pipeline } = joinBuilder({
       relationships: {
         zip567: {
           from: 'posts-collection',
           field: 'posts',
-          pipeline: [
-            { $or: [{ title: { $eq: 'hello' } }, { quux987_labels_some: { $eq: true } }] },
-          ],
+          matchTerm: { $or: [{ title: { $eq: 'hello' } }, { quux987_labels_some: { $eq: true } }] },
+          postJoinPipeline: [],
           postQueryMutation: jest.fn(),
           many: true,
           relationships: {
             quux987: {
               from: 'labels-collection',
               field: 'labels',
-              pipeline: [{ name: { $eq: 'foo' } }],
+              matchTerm: { name: { $eq: 'foo' } },
+              postJoinPipeline: [],
               postQueryMutation: jest.fn(),
               many: true,
+              relationships: {},
             },
           },
         },
       },
-      pipeline: [
-        {
-          $or: [
-            { name: { $eq: 'foobar' } },
-            { age: { $eq: 23 } },
-            { zip567_posts_every: { $eq: true } },
-          ],
-        },
-      ],
+      matchTerm: {
+        $or: [
+          { name: { $eq: 'foobar' } },
+          { age: { $eq: 23 } },
+          { zip567_posts_every: { $eq: true } },
+        ],
+      },
+      postJoinPipeline: [],
     });
 
-    expect(joinQuery).toMatchObject([
+    expect(pipeline).toMatchObject([
       {
         $lookup: {
           from: 'posts-collection',
           as: 'zip567_posts',
           let: { zip567_posts_ids: '$posts' },
           pipeline: [
+            { $match: { $expr: { $in: ['$_id', '$$zip567_posts_ids'] } } },
             {
               $lookup: {
                 from: 'labels-collection',
                 as: 'quux987_labels',
                 let: { quux987_labels_ids: '$labels' },
                 pipeline: [
+                  { $match: { $expr: { $in: ['$_id', '$$quux987_labels_ids'] } } },
                   {
-                    $match: {
-                      $and: [
-                        { $expr: { $in: ['$_id', '$$quux987_labels_ids'] } },
-                        { name: { $eq: 'foo' } },
-                      ],
-                    },
+                    $match: { name: { $eq: 'foo' } },
                   },
                   { $addFields: { id: '$_id' } },
                 ],
@@ -556,12 +550,7 @@ describe('join builder', () => {
             },
             {
               $match: {
-                $and: [
-                  { $expr: { $in: ['$_id', '$$zip567_posts_ids'] } },
-                  {
-                    $or: [{ title: { $eq: 'hello' } }, { quux987_labels_some: { $eq: true } }],
-                  },
-                ],
+                $or: [{ title: { $eq: 'hello' } }, { quux987_labels_some: { $eq: true } }],
               },
             },
             { $addFields: { id: '$_id' } },
@@ -605,58 +594,55 @@ describe('join builder', () => {
       }
     */
 
-    const { joinQuery } = joinBuilder({
+    const { pipeline } = joinBuilder({
       relationships: {
         zip567: {
           from: 'posts-collection',
           field: 'posts',
-          pipeline: [
-            { $or: [{ title: { $eq: 'hello' } }, { quux987_labels_some: { $eq: true } }] },
-          ],
+          matchTerm: { $or: [{ title: { $eq: 'hello' } }, { quux987_labels_some: { $eq: true } }] },
+          postJoinPipeline: [],
           postQueryMutation: jest.fn(),
           many: true,
           relationships: {
             quux987: {
               from: 'labels-collection',
               field: 'labels',
-              pipeline: [{ name: { $eq: 'foo' } }],
+              matchTerm: { name: { $eq: 'foo' } },
+              postJoinPipeline: [],
               postQueryMutation: jest.fn(),
               many: true,
+              relationships: {},
             },
           },
         },
       },
-      pipeline: [
-        {
-          $and: [
-            { name: { $eq: 'foobar' } },
-            { age: { $eq: 23 } },
-            { zip567_posts_every: { $eq: true } },
-          ],
-        },
-      ],
+      matchTerm: {
+        $and: [
+          { name: { $eq: 'foobar' } },
+          { age: { $eq: 23 } },
+          { zip567_posts_every: { $eq: true } },
+        ],
+      },
+      postJoinPipeline: [],
     });
 
-    expect(joinQuery).toMatchObject([
+    expect(pipeline).toMatchObject([
       {
         $lookup: {
           from: 'posts-collection',
           as: 'zip567_posts',
           let: { zip567_posts_ids: '$posts' },
           pipeline: [
+            { $match: { $expr: { $in: ['$_id', '$$zip567_posts_ids'] } } },
             {
               $lookup: {
                 from: 'labels-collection',
                 as: 'quux987_labels',
                 let: { quux987_labels_ids: '$labels' },
                 pipeline: [
+                  { $match: { $expr: { $in: ['$_id', '$$quux987_labels_ids'] } } },
                   {
-                    $match: {
-                      $and: [
-                        { $expr: { $in: ['$_id', '$$quux987_labels_ids'] } },
-                        { name: { $eq: 'foo' } },
-                      ],
-                    },
+                    $match: { name: { $eq: 'foo' } },
                   },
                   { $addFields: { id: '$_id' } },
                 ],
@@ -671,12 +657,7 @@ describe('join builder', () => {
             },
             {
               $match: {
-                $and: [
-                  { $expr: { $in: ['$_id', '$$zip567_posts_ids'] } },
-                  {
-                    $or: [{ title: { $eq: 'hello' } }, { quux987_labels_some: { $eq: true } }],
-                  },
-                ],
+                $or: [{ title: { $eq: 'hello' } }, { quux987_labels_some: { $eq: true } }],
               },
             },
             { $addFields: { id: '$_id' } },
@@ -720,59 +701,56 @@ describe('join builder', () => {
       }
     */
 
-    const { joinQuery } = joinBuilder({
+    const { pipeline } = joinBuilder({
       relationships: {
         zip567: {
           from: 'posts-collection',
           field: 'posts',
-          pipeline: [
-            { $and: [{ title: { $eq: 'hello' } }, { quux987_labels_some: { $eq: true } }] },
-          ],
+          matchTerm: {
+            $and: [{ title: { $eq: 'hello' } }, { quux987_labels_some: { $eq: true } }],
+          },
+          postJoinPipeline: [],
           postQueryMutation: jest.fn(),
           many: true,
           relationships: {
             quux987: {
               from: 'labels-collection',
               field: 'labels',
-              pipeline: [{ name: { $eq: 'foo' } }],
+              matchTerm: { name: { $eq: 'foo' } },
+              postJoinPipeline: [],
               postQueryMutation: jest.fn(),
               many: true,
+              relationships: {},
             },
           },
         },
       },
-      pipeline: [
-        {
-          $or: [
-            { name: { $eq: 'foobar' } },
-            { age: { $eq: 23 } },
-            { zip567_posts_every: { $eq: true } },
-          ],
-        },
-      ],
+      matchTerm: {
+        $or: [
+          { name: { $eq: 'foobar' } },
+          { age: { $eq: 23 } },
+          { zip567_posts_every: { $eq: true } },
+        ],
+      },
+      postJoinPipeline: [],
     });
 
-    expect(joinQuery).toMatchObject([
+    expect(pipeline).toMatchObject([
       {
         $lookup: {
           from: 'posts-collection',
           as: 'zip567_posts',
           let: { zip567_posts_ids: '$posts' },
           pipeline: [
+            { $match: { $expr: { $in: ['$_id', '$$zip567_posts_ids'] } } },
             {
               $lookup: {
                 from: 'labels-collection',
                 as: 'quux987_labels',
                 let: { quux987_labels_ids: '$labels' },
                 pipeline: [
-                  {
-                    $match: {
-                      $and: [
-                        { $expr: { $in: ['$_id', '$$quux987_labels_ids'] } },
-                        { name: { $eq: 'foo' } },
-                      ],
-                    },
-                  },
+                  { $match: { $expr: { $in: ['$_id', '$$quux987_labels_ids'] } } },
+                  { $match: { name: { $eq: 'foo' } } },
                   { $addFields: { id: '$_id' } },
                 ],
               },
@@ -786,12 +764,7 @@ describe('join builder', () => {
             },
             {
               $match: {
-                $and: [
-                  { $expr: { $in: ['$_id', '$$zip567_posts_ids'] } },
-                  {
-                    $and: [{ title: { $eq: 'hello' } }, { quux987_labels_some: { $eq: true } }],
-                  },
-                ],
+                $and: [{ title: { $eq: 'hello' } }, { quux987_labels_some: { $eq: true } }],
               },
             },
             { $addFields: { id: '$_id' } },
@@ -832,18 +805,20 @@ describe('join builder', () => {
     const mutationResult = {};
     const postQueryMutation = jest.fn(() => mutationResult);
 
-    const { mutator } = joinBuilder({
+    const { postQueryMutations } = joinBuilder({
       relationships: {
         zip567: {
           from: 'posts-collection',
           field: 'posts',
-          pipeline: [{ title: { $eq: 'hello' } }],
-          match: [{ zip567_posts_every: { $eq: true } }],
+          matchTerm: { title: { $eq: 'hello' } },
+          postJoinPipeline: [],
           postQueryMutation,
           many: true,
+          relationships: {},
         },
       },
-      pipeline: [{ age: { $eq: 23 } }],
+      matchTerm: { $and: [{ age: { $eq: 23 } }, { zip567_posts_every: { $eq: true } }] },
+      postJoinPipeline: [],
     });
 
     /*
@@ -884,7 +859,7 @@ describe('join builder', () => {
 
     const mockQueryResult = [mockResult1, mockResult2];
 
-    const mutatedResult = mutator(mockQueryResult);
+    const mutatedResult = postQueryMutations(mockQueryResult);
 
     // (parentValue, keyOfRelationship, rootObject, path)
     expect(postQueryMutation).toHaveBeenCalledTimes(2);
@@ -941,38 +916,40 @@ describe('join builder', () => {
       }
 
     */
-    const { mutator } = joinBuilder({
+    const { postQueryMutations } = joinBuilder({
       relationships: {
         abc123: {
           from: 'posts-collection',
           field: 'posts',
-          pipeline: [],
-          match: [{ abc123_posts_some: { $eq: true } }],
+          postJoinPipeline: [],
+          matchTerm: { def456_tags_some: { $eq: true } },
           postQueryMutation: abc123_postQueryMutation,
           many: true,
           relationships: {
             def456: {
               from: 'tags-collection',
               field: 'tags',
-              pipeline: [],
-              match: [{ def456_tags_some: { $eq: true } }],
+              postJoinPipeline: [],
+              matchTerm: { xyz890_posts_every: { $eq: true } },
               postQueryMutation: def456_postQueryMutation,
               many: true,
               relationships: {
                 xyz890: {
                   from: 'posts-collection',
                   field: 'posts',
-                  pipeline: [{ published: { $eq: true } }],
-                  match: [{ xyz890_posts_every: { $eq: true } }],
+                  matchTerm: { published: { $eq: true } },
+                  postJoinPipeline: [],
                   postQueryMutation: xyz890_postQueryMutation,
                   many: true,
+                  relationships: {},
                 },
               },
             },
           },
         },
       },
-      pipeline: [],
+      matchTerm: { abc123_posts_some: { $eq: true } },
+      postJoinPipeline: [],
     });
 
     /*
@@ -1017,7 +994,7 @@ describe('join builder', () => {
       },
     ];
 
-    const mutatedResult = mutator(mockQueryResult);
+    const mutatedResult = postQueryMutations(mockQueryResult);
 
     // xyz890 relationship
     expect(xyz890_postQueryMutation).toHaveBeenCalledTimes(3);