+++ packages/adapter-mongoose/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @voussoir/adapter-mongoose
 
+## 0.3.1
+- [patch] Updated dependencies [74af97e](74af97e)
+  - @voussoir/mongo-join-builder@0.2.0
+
+## 0.3.0
+
+- [patch] Bump all packages for Babel config fixes [d51c833](d51c833)
+- [minor] Support unique field constraint for mongoose adapter [750a83e](750a83e)
+- [patch] Updated dependencies [9c75136](9c75136)
+  - @voussoir/core@0.3.0
+  - @voussoir/mongo-join-builder@0.1.3
+  - @voussoir/utils@0.2.0
+
 ## 0.2.0
 
 - [minor] Add missing dependencies for which the mono-repo was hiding that they were missing [fed0cdc](fed0cdc)