+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @voussoir/icons
 
+## 0.1.3
+
+- [patch] Bump all packages for Babel config fixes [d51c833](d51c833)
+
 ## 0.1.2
 
 - [patch] Rename readme files [a8b995e](a8b995e)